SummaryService in its own directory
