+++ src/app/summary/pages/summary/summary.component.ts
@@ -1,7 +1,7 @@
 import {ChangeDetectionStrategy, Component, OnInit} from '@angular/core';
 import {Observable} from "rxjs";
 import {Project} from "../../../projects/project";
-import {ProjectCount, SummaryService} from "../../summary.service";
+import {ProjectCount, SummaryService} from "../../services/summary.service";
 import {ProjectsService} from "../../../projects/services/projects.service";
 import {HeadingService} from "../../../services/heading.service";
