+++ src/app/summary/services/summary.service.spec.ts
@@ -1,42 +1,13 @@
-import {async, TestBed} from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
 import {SummaryService} from './summary.service';
-import {ProjectsService} from "../projects/services/projects.service";
+import {ProjectsService} from "../../projects/services/projects.service";
 import {Observable, of} from "rxjs";
 import {HttpClientTestingModule} from "@angular/common/http/testing";
-import {Project} from "../projects/project";
+import {Project} from "../../projects/project";
 import {delay} from "rxjs/operators";
+import {makeDummyProjects} from "./projects.data";
 
-
-function makeProject() {
-  return {
-    addedToIndex: "",
-    arrayExpressAccessions: [],
-    authors: [],
-    date: "",
-    dcpUrl: "",
-    egaDatasetsAccessions: [],
-    egaStudiesAccessions: [],
-    enaAccessions: [],
-    geoAccessions: [],
-    publications: [],
-    title: "",
-    uuid: "",
-    organs: [],
-    technologies: [],
-    cellCount: 0
-  };
-}
-
-function makeDummyProjects() {
-  let dummyProject: Project = makeProject();
-
-  let projects: Project[] = [
-    {...dummyProject, organs: ['lungs', 'brain'], technologies: ['t1', 't2'], cellCount: 200},
-    {...dummyProject, organs: ['kidney', 'lungs'], technologies: ['t1', 't3'], cellCount: 300}
-  ];
-  return projects;
-}
 
 class MockProjectsService {
   getAllProjects() :Observable<Project[]>{
